+++ examples/tip/prod/prop_07.scala
@@ -22,25 +22,24 @@ object prop_07 extends TacticsProof {
 
   val sequent =
     hols"""
-        def_head: ∀x0 ∀x1 (head(cons(x0:sk, x1:list): list): sk) = x0,
-        def_tail: ∀x0 ∀x1 (tail(cons(x0:sk, x1:list): list): list) = x1,
-        def_p: ∀x0 (p(S(x0:Nat): Nat): Nat) = x0,
-        def_qrev_0: ∀y (qrev(nil:list, y:list): list) = y,
-        def_qrev_1: ∀z ∀xs ∀y (qrev(cons(z:sk, xs:list): list, y:list): list) = qrev(xs, cons(z, y)),
-        def_plus_0: ∀y (plus(#c(Z: Nat), y:Nat): Nat) = y,
-        def_plus_1: ∀z ∀y (plus(S(z:Nat): Nat, y:Nat): Nat) = S(plus(z, y)),
-        def_length_0: (length(nil:list): Nat) = #c(Z: Nat),
-        def_length_1: ∀y ∀xs (length(cons(y:sk, xs:list): list): Nat) = S(length(xs)),
-        constr_inj_0: ∀y0 ∀y1 ¬(nil:list) = cons(y0:sk, y1:list),
-        constr_inj_1: ∀y0 ¬#c(Z: Nat) = S(y0:Nat)
+        def_head: ∀x0 ∀x1 head(cons(x0, x1)) = x0,
+        def_tail: ∀x0 ∀x1 tail(cons(x0, x1)) = x1,
+        def_p: ∀x0 p(S(x0)) = x0,
+        def_qrev_0: ∀y qrev(nil, y) = y,
+        def_qrev_1: ∀z ∀xs ∀y qrev(cons(z, xs), y) = qrev(xs, cons(z, y)),
+        def_plus_0: ∀y plus(Z, y) = y,
+        def_plus_1: ∀z ∀y plus(S(z), y) = S(plus(z, y)),
+        def_length_0: length(nil) = Z,
+        def_length_1: ∀y ∀xs length(cons(y, xs)) = S(length(xs)),
+        constr_inj_0: ∀y0 ∀y1 ¬nil = cons(y0, y1),
+        constr_inj_1: ∀y0 ¬Z = S(y0)
         :-
-        goal: ∀x ∀y (length(qrev(x:list, y:list): list): Nat) = plus(length(x), length(y))
+        goal: ∀x ∀y length(qrev(x, y)) = plus(length(x), length(y))
   """
 
-  val lem_1 = (
-    ( "ap1" -> hof"∀y plus(Z, y) = y" ) +:
+  val lem_1 = ( "ap1" -> hof"∀y plus(Z, y) = y" ) +:
     ( "ap2" -> hof"∀z ∀y plus(S(z), y) = S(plus(z, y))" ) +:
-    Sequent() :+ ( "lem_1" -> hof"∀x ∀y plus(x,S(y)) = S(plus(x,y))" ) )
+    Sequent() :+ ( "lem_1" -> hof"∀x ∀y plus(x,S(y)) = S(plus(x,y))" )
 
   val lem_1_proof = AnalyticInductionProver.singleInduction( lem_1, hov"x:Nat" )
 
@@ -52,5 +51,14 @@ object prop_07 extends TacticsProof {
     cut( "lem_1", hof"∀x ∀y plus(x,S(y)) = S(plus(x,y))" )
     insert( lem_1_proof )
     insert( cut_lem_proof )
+  }
+
+  val proof2 = Lemma( sequent ) {
+    cut( "l", hof"""
+        !x!y (length(qrev x y) = plus(length x, length y) &
+              plus(length x, S(length y)) = S(plus(length x, length y)))
+      """ ) right escrgt
+    forget( "goal" ); allR( hov"x:list" )
+    induction( hov"x:list" ) onAll escrgt
   }
 }