+++ RELEASE-NOTES.md
@@ -1,6 +1,10 @@
 # Release notes for GAPT
 
-## Version 2.11 (unreleased)
+## Version 2.12 (unreleased)
+
+* SMTInterpol interface
+
+## Version 2.11 (released on 2018-04-30)
 
 * Deskolemization of proofs with equational reasoning
 * Root package was renamed from `at.logic.gapt` to just `gapt`